refactor(launchconfiguration): Move to autoscaling service
