+++ internal/service/autoscaling/launch_configuration.go
@@ -1,4 +1,4 @@
-package launchconfiguration
+package autoscaling
 
 import (
 	"fmt"
