+++ internal/service/autoscaling/launch_configuration_data_source.go
@@ -1,4 +1,4 @@
-package launchconfiguration
+package autoscaling
 
 import (
 	"github.com/NaverCloudPlatform/ncloud-sdk-go-v2/ncloud"
