+++ internal/service/autoscaling/launch_configuration_data_source_test.go
@@ -1,4 +1,4 @@
-package launchconfiguration_test
+package autoscaling_test
 
 import (
 	"fmt"
